+++ packages/atomic/src/Button/button.utils.ts
@@ -1,8 +1,12 @@
-export const button = (disabled = false) => {
-  return {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    cursor: disabled ? 'not-allowed' : 'pointer',
-  };
-};
+import { RCSSProperties } from '@unstyled-ui/core';
+
+export const button = (css: RCSSProperties = {}) => ({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  cursor: 'pointer',
+  '&[disabled]': {
+    cursor: 'not-allowed',
+  },
+  ...css,
+});