+++ packages/core/src/types.ts
@@ -22,5 +22,5 @@ export type BaseProps<
   Attr extends React.HTMLAttributes<HTMLElement> = React.HTMLAttributes<HTMLElement>
 > = VariantProps<any> & {
   css?: RCSSProps;
-  as?: keyof JSX.IntrinsicElements;
+  as?: keyof JSX.IntrinsicElements | string;
 } & Attr;