+++ packages/layout/src/Abs/Abs.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 
 export const Abs: React.FC<BaseProps> = props => {
   const { css, ...restProps } = props;
-  //@ts-ignore
   return (
+    //@ts-ignore
     <Box as="u-abs" css={{ ...css, position: 'absolute' }} {...restProps} />
   );
 };