+++ packages/atomic/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 1.0.3
 
+## 1.0.8
+
+### Patch Changes
+
+- d146218: form and item component
+- Updated dependencies [d146218]
+  - @unstyled-ui/core@1.0.7
+  - @unstyled-ui/stitches@1.0.9
+
 ## 1.0.7
 
 ### Patch Changes