+++ packages/stitches/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 1.0.2
 
+## 1.0.9
+
+### Patch Changes
+
+- d146218: form and item component
+
 ## 1.0.8
 
 ### Patch Changes